+++ lib/hmmlearn/_utils.py
@@ -1,7 +1,5 @@
 """Private utilities."""
 
-import warnings
-
 import numpy as np
 from sklearn.utils.validation import NotFittedError
 
@@ -12,14 +10,9 @@ def split_X_lengths(X, lengths):
     else:
         cs = np.cumsum(lengths)
         n_samples = len(X)
-        if cs[-1] > n_samples:
+        if cs[-1] != n_samples:
             raise ValueError(
-                f"more than {n_samples} samples in lengths array {lengths}")
-        elif cs[-1] != n_samples:
-            warnings.warn(
-                f"less that {n_samples} samples in lengths array {lengths}; "
-                f"support for silently dropping samples is deprecated and "
-                f"will be removed", DeprecationWarning, stacklevel=3)
+                f"lengths array {lengths} doesn't sum to {n_samples} samples")
         return np.split(X, cs)[:-1]